app/requirements.txt
absl-py==2.1.0
altair==5.2.0
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
comm==0.2.1
comtypes==1.4.1
contourpy==1.2.0
cv==1.0.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.49.0
fqdn==1.5.1
fsspec==2024.6.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.19.0
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.24.6
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.12.3
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.26
jaxlib==0.4.26
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.13
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.23.5
oauthlib==3.2.2
openai==1.33.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==23.2
pandas==1.5.1
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipenv==2023.12.1
pipreqs==0.5.0
platformdirs==4.2.0
playsound==1.3.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2022.6
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2022.10.31
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.4
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
stack-data==0.6.3
streamlit==1.31.1
sympy==1.13.2
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
thop==0.1.1.post2209072238
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.64.1
traitlets==5.14.1
transformers==4.44.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.0
uri-template==1.3.0
urllib3==2.2.1
validators==0.22.0
virtualenv==20.25.1
watchdog==4.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.17.0
